pyproject.toml
[project]
name = "multi-modal-rag-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Oleksii Postovyi",email = "a.postovoy918@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "scrapy (>=2.13.1,<3.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langgraph (>=0.4.7,<0.5.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "langchain-chroma (>=0.2.4,<0.3.0)",
    "langchain-openai (>=0.3.18,<0.4.0)",
    "fastapi (==0.115.9)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "ruff (>=0.11.12,<0.12.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "streamlit (>=1.45.1,<2.0.0)",
    "deepeval (>=3.0.2,<4.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "pytest-asyncio (>=1.0.0,<2.0.0)",
    "langsmith (>=0.3.43,<0.4.0)",
]
[tool.poetry]
name = "multi-modal-rag-bot"
version = "0.1.0"
description = "A multi-modal RAG bot application"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
click = "*"
kubernetes = "*"
lxml = "*"
numpy = "*"
protobuf = "*"
pyyaml = "*"
requests = "*"
scrapy = "*"
six = "*"
twisted = "*"
sympy = "*"
wrapt = "*"
sqlalchemy = "*"
chromadb = "*"
langchain = "*"
langchain-chroma = "*"
langchain-core = "*"
langchain-openai = "*"
langgraph = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
